refactor: tokens module structure
